feat(lazy): move plugin setups to lua/
