+++ init.lua
@@ -1,1 +1,3 @@
+vim.g.mapleader = " "
+vim.g.localleader = "\\"
 require("yoolayn.lazy")
