+++ lua/yoolayn/plugins/dap.lua
@@ -1,0 +1,38 @@
+return {
+	"mfussenegger/nvim-dap",
+	dependencies = {
+		"rcarriga/nvim-dap-ui",
+		"theHamsta/nvim-dap-virtual-text",
+		"leoluz/nvim-dap-go",
+		"mfussenegger/nvim-dap-python",
+	},
+	config = function()
+		local dap = require("dap")
+		require("nvim-dap-virtual-text").setup()
+		require("dapui").setup()
+		require("dap-go").setup()
+
+		local dap, dapui = require("dap"), require("dapui")
+		dap.listeners.after.event_initialized["dapui_config"] = function()
+			dapui.open()
+		end
+		dap.listeners.before.event_terminated["dapui_config"] = function()
+			dapui.close()
+		end
+		dap.listeners.before.event_exited["dapui_config"] = function()
+			dapui.close()
+		end
+
+		dap.configurations.scala = {
+			{
+				type = "scala",
+				request = "launch",
+				name = "Run with arg and env file",
+				metals = {
+				},
+			}
+		}
+
+		vim.fn.sign_define("DapStopped", { text = "ඞ", texthl = "Error" })
+	end,
+}
