+++ lua/yoolayn/plugins/noconfig.lua
@@ -1,0 +1,9 @@
+return {
+	"cohama/lexima.vim",
+	"stefandtw/quickfix-reflector.vim",
+	"kristijanhusak/vim-dadbod-ui",
+	"godlygeek/tabular",
+	"dstein64/vim-startuptime",
+	"ThePrimeagen/vim-be-good",
+	"anuvyklack/hydra.nvim" ,
+}
