+++ lua/yoolayn/plugins/tpope.lua
@@ -1,0 +1,8 @@
+return {
+	"tpope/vim-dadbod",
+	"tpope/vim-dispatch",
+	"tpope/vim-eunuch",
+	"tpope/vim-fugitive",
+	"tpope/vim-repeat",
+	"tpope/vim-surround",
+}
